Started to add unit tests for the new spectrum class

+++ CSMSL/IO/Agilent/AgilentDDirectory.cs
@@ -148,7 +148,7 @@ namespace CSMSL.IO.Agilent
             return new MzRange(spectrum.MeasuredMassRange.Start, spectrum.MeasuredMassRange.End);
         }
 
-        public override short GetPrecusorCharge(int spectrumNumber, int msnOrder = 2)
+        public override int GetPrecusorCharge(int spectrumNumber, int msnOrder = 2)
         {
             IBDASpecData spectrum = _msdr.GetSpectrum(spectrumNumber - 1);
             int precursor_charge;
